Implements better versions, projects
This commit introduces a bunch of changes that make versions and
projects more useful as models, and that help normalize the pages model
by moving a few columns up into the version model.

The rake tasks around file import are dramatically cut down as a result,
and the version model picks up the ability to import its own files.